Fixed bootstrap issues with the views. Fixed ng-app stuff
I added in a concurrent version of bootstrap that we can use to design the website. The ng-app was in the wrong place and it was screwing with ng-view. So I put ng-app in the html tag at the top so that we declare all of index as part of the application.